requirements.txt
Django==1.9
djangorestframework==3.7.3
simplejson==3.8.0
virtualenv==13.1.2
wheel==0.26.0